docs(script): 변경된 소스 반영 로직 수정
- 기존: 백업 -> 반영 -> 백업본 삭제
- 수정: 기존 백업본 삭제 -> 현재 소스 백업 -> 반영

- 반영 완료 후 백업본을 삭제하지 않고 다음 반영 전까지 유지함으로써
  반영한 버전에 문제가 있을 경우 백업본을 사용, 빠른 롤백이 가능하도록함

+++ script/deploy.sh
@@ -162,8 +162,14 @@ backup() {
 # app 소스 백업 삭제
 delete_backup() {
   echo -e "${txtcyn}delete backup dir($BUILD_BACKUP_NAME)${txtrst}"
-  rm -rf $APP_PATH$BUILD_BACKUP_NAME
-  check_error
+  if [ -d "$APP_PATH$BUILD_BACKUP_NAME" ]
+  then
+    rm -rf $APP_PATH$BUILD_BACKUP_NAME
+    check_error
+  else
+    echo -e "${txtmgb}cannot find directory $APP_PATH$BUILD_BACKUP_NAME${txtrst}${txtdfb}"
+    echo -e "${txtmgb}skip deleting backup dir...${txtrst}${txtdfb}"
+  fi
 }
 
 # app 소스 build 소스로 덮어씌우기
@@ -204,13 +210,13 @@ check_diff() {
     exit_script 0
   else
     echo -e "${txtmgb}difference detected. start deploy${txtrst}${txtdfb}"
+    delete_backup
     backup
     pull n
     build
     service_stop
     overwrite_app
     service_start $mode
-    delete_backup
   fi
 }
